feat(types): allow extending mapStores suffix
