+++ .gitignore
@@ -7,3 +7,4 @@ coverage.lcov
 dist
 package-lock.json
 .DS_Store
+test-dts/tsconfig.tsbuildinfo
